feat: v2 代码迁移 for tag: v2-20230307

+++ jeecg-module-system/jeecg-system-biz/src/main/java/org/jeecg/modules/message/controller/SysMessageController.java
@@ -1,31 +1,22 @@
 package org.jeecg.modules.message.controller;
 
-import java.util.Arrays;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.poi.ss.formula.functions.T;
 import org.jeecg.common.api.vo.Result;
 import org.jeecg.common.system.base.controller.JeecgController;
 import org.jeecg.common.system.query.QueryGenerator;
 import org.jeecg.modules.message.entity.SysMessage;
 import org.jeecg.modules.message.service.ISysMessageService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
-import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.baomidou.mybatisplus.core.metadata.IPage;
-import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-
-import lombok.extern.slf4j.Slf4j;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.util.Arrays;
 
 /**
  * @Description: 消息
@@ -35,7 +26,7 @@ import lombok.extern.slf4j.Slf4j;
  */
 @Slf4j
 @RestController
-@RequestMapping("/sys/message/sysMessage")
+@RequestMapping("/message/sysMessage")
 public class SysMessageController extends JeecgController<SysMessage, ISysMessageService> {
 	@Autowired
 	private ISysMessageService sysMessageService;
@@ -50,12 +41,17 @@ public class SysMessageController extends JeecgController<SysMessage, ISysMessag
 	 * @return
 	 */
 	@GetMapping(value = "/list")
-	public Result<?> queryPageList(SysMessage sysMessage, @RequestParam(name = "pageNo", defaultValue = "1") Integer pageNo,
-			@RequestParam(name = "pageSize", defaultValue = "10") Integer pageSize, HttpServletRequest req) {
+	public Result<IPage<SysMessage>> queryPageList(SysMessage sysMessage,
+												   @RequestParam(name = "pageNo", defaultValue = "1") Integer pageNo,
+												   @RequestParam(name = "pageSize", defaultValue = "10") Integer pageSize,
+												   HttpServletRequest req) {
+		Result<IPage<SysMessage>> result = new Result<IPage<SysMessage>>();
 		QueryWrapper<SysMessage> queryWrapper = QueryGenerator.initQueryWrapper(sysMessage, req.getParameterMap());
 		Page<SysMessage> page = new Page<SysMessage>(pageNo, pageSize);
 		IPage<SysMessage> pageList = sysMessageService.page(page, queryWrapper);
-        return Result.ok(pageList);
+		result.setSuccess(true);
+		result.setResult(pageList);
+		return result;
 	}
 
 	/**
@@ -65,9 +61,16 @@ public class SysMessageController extends JeecgController<SysMessage, ISysMessag
 	 * @return
 	 */
 	@PostMapping(value = "/add")
-	public Result<?> add(@RequestBody SysMessage sysMessage) {
-		sysMessageService.save(sysMessage);
-		return Result.ok("添加成功！");
+	public Result<T> add(@RequestBody SysMessage sysMessage) {
+		Result<T> result = new Result<T>();
+		try {
+			sysMessageService.save(sysMessage);
+			result.success("添加成功！");
+		} catch (Exception e) {
+			log.info(e.getMessage(), e);
+			result.error500("操作失败");
+		}
+		return result;
 	}
 
 	/**
@@ -77,10 +80,20 @@ public class SysMessageController extends JeecgController<SysMessage, ISysMessag
 	 * @return
 	 */
 	@PutMapping(value = "/edit")
-	public Result<?> edit(@RequestBody SysMessage sysMessage) {	
-		sysMessageService.updateById(sysMessage);
-        return Result.ok("修改成功!");
-
+	public Result<T> edit(@RequestBody SysMessage sysMessage) {
+		Result<T> result = new Result<T>();
+		SysMessage sysMessageEntity = sysMessageService.getById(sysMessage.getId());
+		if (sysMessageEntity == null) {
+			result.error500("未找到对应实体");
+		} else {
+			boolean ok = sysMessageService.updateById(sysMessage);
+			if (ok) {
+				result.success("修改成功!");
+			} else {
+				result.error500("修改失败!");
+			}
+		}
+		return result;
 	}
 
 	/**
@@ -90,9 +103,18 @@ public class SysMessageController extends JeecgController<SysMessage, ISysMessag
 	 * @return
 	 */
 	@DeleteMapping(value = "/delete")
-	public Result<?> delete(@RequestParam(name = "id", required = true) String id) {
-		sysMessageService.removeById(id);
-        return Result.ok("删除成功!");
+	public Result<T> delete(@RequestParam(name = "id", required = true) String id) {
+		Result<T> result = new Result<T>();
+		SysMessage sysMessage = sysMessageService.getById(id);
+		if (sysMessage == null) {
+			result.error500("未找到对应实体");
+		} else {
+			boolean ok = sysMessageService.removeById(id);
+			if (ok) {
+				result.success("删除成功!");
+			}
+		}
+		return result;
 	}
 
 	/**
@@ -102,10 +124,15 @@ public class SysMessageController extends JeecgController<SysMessage, ISysMessag
 	 * @return
 	 */
 	@DeleteMapping(value = "/deleteBatch")
-	public Result<?> deleteBatch(@RequestParam(name = "ids", required = true) String ids) {
-
-		this.sysMessageService.removeByIds(Arrays.asList(ids.split(",")));
-	    return Result.ok("批量删除成功！");
+	public Result<T> deleteBatch(@RequestParam(name = "ids", required = true) String ids) {
+		Result<T> result = new Result<T>();
+		if (ids == null || "".equals(ids.trim())) {
+			result.error500("ids参数不允许为空！");
+		} else {
+			this.sysMessageService.removeByIds(Arrays.asList(ids.split(",")));
+			result.success("删除成功!");
+		}
+		return result;
 	}
 
 	/**
@@ -115,15 +142,23 @@ public class SysMessageController extends JeecgController<SysMessage, ISysMessag
 	 * @return
 	 */
 	@GetMapping(value = "/queryById")
-	public Result<?> queryById(@RequestParam(name = "id", required = true) String id) {
+	public Result<SysMessage> queryById(@RequestParam(name = "id", required = true) String id) {
+		Result<SysMessage> result = new Result<SysMessage>();
 		SysMessage sysMessage = sysMessageService.getById(id);
-		return Result.ok(sysMessage);
+		if (sysMessage == null) {
+			result.error500("未找到对应实体");
+		} else {
+			result.setResult(sysMessage);
+			result.setSuccess(true);
+		}
+		return result;
 	}
 
 	/**
 	 * 导出excel
 	 *
 	 * @param request
+	 * @param response
 	 */
 	@GetMapping(value = "/exportXls")
 	public ModelAndView exportXls(HttpServletRequest request, SysMessage sysMessage) {
