+++ jeecg-module-system/jeecg-system-biz/src/main/java/org/jeecg/modules/message/controller/SysMessageTemplateController.java
@@ -1,38 +1,26 @@
 package org.jeecg.modules.message.controller;
 
-import java.util.Arrays;
-import java.util.Map;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.jeecg.common.api.dto.message.MessageDTO;
+import com.alibaba.fastjson.JSON;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.poi.ss.formula.functions.T;
 import org.jeecg.common.api.vo.Result;
-import org.jeecg.common.system.api.ISysBaseAPI;
 import org.jeecg.common.system.base.controller.JeecgController;
 import org.jeecg.common.system.query.QueryGenerator;
-import org.jeecg.common.util.oConvertUtils;
 import org.jeecg.modules.message.entity.MsgParams;
 import org.jeecg.modules.message.entity.SysMessageTemplate;
 import org.jeecg.modules.message.service.ISysMessageTemplateService;
 import org.jeecg.modules.message.util.PushMsgUtil;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
-import com.alibaba.fastjson.JSON;
-import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.baomidou.mybatisplus.core.metadata.IPage;
-import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-
-import lombok.extern.slf4j.Slf4j;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.util.Arrays;
+import java.util.Map;
 
 /**
  * @Description: 消息模板
@@ -42,15 +30,12 @@ import lombok.extern.slf4j.Slf4j;
  */
 @Slf4j
 @RestController
-@RequestMapping("/sys/message/sysMessageTemplate")
+@RequestMapping("/message/sysMessageTemplate")
 public class SysMessageTemplateController extends JeecgController<SysMessageTemplate, ISysMessageTemplateService> {
 	@Autowired
 	private ISysMessageTemplateService sysMessageTemplateService;
 	@Autowired
 	private PushMsgUtil pushMsgUtil;
-
-	@Autowired
-	private ISysBaseAPI sysBaseApi;
 
 	/**
 	 * 分页列表查询
@@ -62,12 +47,15 @@ public class SysMessageTemplateController extends JeecgController<SysMessageTemp
 	 * @return
 	 */
 	@GetMapping(value = "/list")
-	public Result<?> queryPageList(SysMessageTemplate sysMessageTemplate, @RequestParam(name = "pageNo", defaultValue = "1") Integer pageNo,
+	public Result<IPage<SysMessageTemplate>> queryPageList(SysMessageTemplate sysMessageTemplate, @RequestParam(name = "pageNo", defaultValue = "1") Integer pageNo,
 			@RequestParam(name = "pageSize", defaultValue = "10") Integer pageSize, HttpServletRequest req) {
+		Result<IPage<SysMessageTemplate>> result = new Result<IPage<SysMessageTemplate>>();
 		QueryWrapper<SysMessageTemplate> queryWrapper = QueryGenerator.initQueryWrapper(sysMessageTemplate, req.getParameterMap());
 		Page<SysMessageTemplate> page = new Page<SysMessageTemplate>(pageNo, pageSize);
 		IPage<SysMessageTemplate> pageList = sysMessageTemplateService.page(page, queryWrapper);
-        return Result.ok(pageList);
+		result.setSuccess(true);
+		result.setResult(pageList);
+		return result;
 	}
 
 	/**
@@ -77,9 +65,16 @@ public class SysMessageTemplateController extends JeecgController<SysMessageTemp
 	 * @return
 	 */
 	@PostMapping(value = "/add")
-	public Result<?> add(@RequestBody SysMessageTemplate sysMessageTemplate) {
-		sysMessageTemplateService.save(sysMessageTemplate);
-        return Result.ok("添加成功！");
+	public Result<T> add(@RequestBody SysMessageTemplate sysMessageTemplate) {
+		Result<T> result = new Result<T>();
+		try {
+			sysMessageTemplateService.save(sysMessageTemplate);
+			result.success("添加成功！");
+		} catch (Exception e) {
+			log.info(e.getMessage(), e);
+			result.error500("操作失败");
+		}
+		return result;
 	}
 
 	/**
@@ -89,9 +84,21 @@ public class SysMessageTemplateController extends JeecgController<SysMessageTemp
 	 * @return
 	 */
 	@PutMapping(value = "/edit")
-	public Result<?> edit(@RequestBody SysMessageTemplate sysMessageTemplate) {
-		sysMessageTemplateService.updateById(sysMessageTemplate);
-        return Result.ok("更新成功！");
+	public Result<T> edit(@RequestBody SysMessageTemplate sysMessageTemplate) {
+		Result<T> result = new Result<T>();
+		SysMessageTemplate sysMessageTemplateEntity = sysMessageTemplateService.getById(sysMessageTemplate.getId());
+		if (sysMessageTemplateEntity == null) {
+			result.error500("未找到对应实体");
+		} else {
+			boolean ok = sysMessageTemplateService.updateById(sysMessageTemplate);
+			if (ok) {
+				result.success("修改成功!");
+			} else {
+				result.error500("修改失败!");
+			}
+		}
+
+		return result;
 	}
 
 	/**
@@ -101,9 +108,19 @@ public class SysMessageTemplateController extends JeecgController<SysMessageTemp
 	 * @return
 	 */
 	@DeleteMapping(value = "/delete")
-	public Result<?> delete(@RequestParam(name = "id", required = true) String id) {
-		sysMessageTemplateService.removeById(id);
-        return Result.ok("删除成功!");
+	public Result<T> delete(@RequestParam(name = "id", required = true) String id) {
+		Result<T> result = new Result<T>();
+		SysMessageTemplate sysMessageTemplate = sysMessageTemplateService.getById(id);
+		if (sysMessageTemplate == null) {
+			result.error500("未找到对应实体");
+		} else {
+			boolean ok = sysMessageTemplateService.removeById(id);
+			if (ok) {
+				result.success("删除成功!");
+			}
+		}
+
+		return result;
 	}
 
 	/**
@@ -113,9 +130,15 @@ public class SysMessageTemplateController extends JeecgController<SysMessageTemp
 	 * @return
 	 */
 	@DeleteMapping(value = "/deleteBatch")
-	public Result<?> deleteBatch(@RequestParam(name = "ids", required = true) String ids) {
-		this.sysMessageTemplateService.removeByIds(Arrays.asList(ids.split(",")));
-        return Result.ok("批量删除成功！");
+	public Result<T> deleteBatch(@RequestParam(name = "ids", required = true) String ids) {
+		Result<T> result = new Result<T>();
+		if (ids == null || "".equals(ids.trim())) {
+			result.error500("ids参数不允许为空！");
+		} else {
+			this.sysMessageTemplateService.removeByIds(Arrays.asList(ids.split(",")));
+			result.success("删除成功!");
+		}
+		return result;
 	}
 
 	/**
@@ -125,15 +148,23 @@ public class SysMessageTemplateController extends JeecgController<SysMessageTemp
 	 * @return
 	 */
 	@GetMapping(value = "/queryById")
-	public Result<?> queryById(@RequestParam(name = "id", required = true) String id) {
+	public Result<SysMessageTemplate> queryById(@RequestParam(name = "id", required = true) String id) {
+		Result<SysMessageTemplate> result = new Result<SysMessageTemplate>();
 		SysMessageTemplate sysMessageTemplate = sysMessageTemplateService.getById(id);
-        return Result.ok(sysMessageTemplate);
+		if (sysMessageTemplate == null) {
+			result.error500("未找到对应实体");
+		} else {
+			result.setResult(sysMessageTemplate);
+			result.setSuccess(true);
+		}
+		return result;
 	}
 
 	/**
 	 * 导出excel
 	 *
 	 * @param request
+	 * @param response
 	 */
 	@GetMapping(value = "/exportXls")
 	public ModelAndView exportXls(HttpServletRequest request,SysMessageTemplate sysMessageTemplate) {
@@ -158,23 +189,19 @@ public class SysMessageTemplateController extends JeecgController<SysMessageTemp
 	@PostMapping(value = "/sendMsg")
 	public Result<SysMessageTemplate> sendMessage(@RequestBody MsgParams msgParams) {
 		Result<SysMessageTemplate> result = new Result<SysMessageTemplate>();
+		Map<String, String> map = null;
 		try {
-			MessageDTO md = new MessageDTO();
-			md.setToAll(false);
-			md.setTitle("消息发送测试");
-			md.setTemplateCode(msgParams.getTemplateCode());
-			md.setToUser(msgParams.getReceiver());
-			md.setType(msgParams.getMsgType());
-			String testData = msgParams.getTestData();
-			if(oConvertUtils.isNotEmpty(testData)){
-				Map<String, Object> data = JSON.parseObject(testData, Map.class);
-				md.setData(data);
-			}
-			sysBaseApi.sendTemplateMessage(md);
-			return result.success("消息发送成功！");
+			map = (Map<String, String>) JSON.parse(msgParams.getTestData());
 		} catch (Exception e) {
-			log.error("发送消息出错", e.getMessage());
-			return result.error500("发送消息出错！");
+			result.error500("解析Json出错！");
+			return result;
 		}
+		boolean is_sendSuccess = pushMsgUtil.sendMessage(msgParams.getMsgType(), msgParams.getTemplateCode(), map, msgParams.getReceiver());
+		if (is_sendSuccess) {
+			result.success("发送消息任务添加成功！");
+		} else {
+			result.error500("发送消息任务添加失败！");
+		}
+		return result;
 	}
 }
